pyproject.toml
[project]
name = "wtr-pop-rr-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "napalm>=5.1.0",
    "netmiko>=4.6.0",
    "nornir>=3.5.0",
    "nornir-netbox>=0.3.0",
    "pyeapi>=1.0.4",
    "pynetbox>=7.5.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scrapli-netconf>=2025.1.30",
    "scrapli[https]>=2025.1.30",
    "xmltodict>=1.0.2",
]
